moved arg parser to main to group user input functions
