+++ pytui/core.py
@@ -1,26 +1,9 @@
-import argparse
 import subprocess  # noqa: S404
 from typing import Any
 
 from parse import findall
 
 from pytui.backends.stackoverflow import StackOverflowFinder
-
-
-def parse_arguments() -> None:
-    """Parse arguments and store them in pytui.arguments.args"""
-    parser = argparse.ArgumentParser()
-    parser.add_argument("-n", "--no-display",
-                        action='store_true',
-                        default=False,
-                        help="Run without display")
-    parser.add_argument("-c", "--copy-error",
-                        action='store_true',
-                        default=False,
-                        help="Copy error to clipboard")
-
-    flags, args = parser.parse_known_args()
-    return vars(flags), args
 
 
 def run_and_get_stderr(args: list[str]) -> str:
